refactor(user): delete AvatarUrlChangeUseCase
